refactor: 정규식 관련 유틸 함수 불러오기 수정
